Compiles music code into audio.
Compiles music code into audio. The resulting composition is played back
and saved to a file. Main point of including this is to demonstrate
ability to stream audio using OpenAL. There are still lots of disabled
features and bugs. One example code is included, but music coding
instructions are not yet.